Fixes 'de-shift' issue on left shift
When the Keyboard library calls write with a symbol that requires
the shift button to print, the left shift is used in that call. As
part of that call the (left) shift release is also called so now
both shift button use the right shift code to avoid shift being
disengaged while still physically held.